chore: add fluent camera API, simplify grid plane configuration, and standardize 3D method naming conventions
- Add fluent Camera setters (set_azimuth/elevation/distance/fov/target) with automatic orbit update
- Replace set_grid_planes(int) with grid_planes(GridPlane) accepting enum directly
- Add bitwise operators (&, ~) for GridPlane enum to complement existing | operator
- Change default grid_planes from XY-only to All for better out-of-box experience
- Rename Axes3D setters to match fluent style

+++ src/ui/workspace.cpp
@@ -648,7 +648,7 @@ WorkspaceData Workspace::capture(const std::vector<Figure*>& figures,
                 a3.z_min = zlim.min; a3.z_max = zlim.max;
                 a3.z_label = ax3d->get_zlabel();
                 a3.camera_state = ax3d->camera().serialize();
-                a3.grid_planes = ax3d->grid_planes();
+                a3.grid_planes = static_cast<int>(ax3d->grid_planes());
                 a3.show_bounding_box = ax3d->show_bounding_box();
                 a3.lighting_enabled = ax3d->lighting_enabled();
                 auto ld = ax3d->light_dir();
